Renamed KitConfig to ConnectConfig
