+++ packages/connect/src/contexts/ConnectConfig.ts
@@ -1,0 +1,7 @@
+'use client'
+
+import { ConnectConfig } from '../types'
+
+import { createGenericContext } from './genericContext'
+
+export const [useConnectConfigContext, ConnectConfigContextProvider] = createGenericContext<ConnectConfig>()
