--- packages/connect/src/contexts/KitConfig.ts
@@ -1,7 +1,0 @@
-'use client'
-
-import { KitConfig } from '../types'
-
-import { createGenericContext } from './genericContext'
-
-export const [useKitConfig, KitConfigContextProvider] = createGenericContext<KitConfig>()
