+++ packages/connect/src/contexts/WalletConfig.ts
